+++ lib/ui/screen/sight_card.dart
@@ -67,7 +67,6 @@ class SightCard extends StatelessWidget {
                     maxWidth: 250,
                   ),
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       const SizedBox(height: 16),
